+++ src/main/java/ua/everybuy/routing/controller/UserController.java
@@ -9,10 +9,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import ua.everybuy.buisnesslogic.service.UserService;
-import ua.everybuy.routing.model.model.dto.UserDto;
-import ua.everybuy.routing.model.model.response.ErrorResponse;
-import ua.everybuy.routing.model.model.response.FullNameResponse;
-import ua.everybuy.routing.model.model.response.StatusResponse;
+import ua.everybuy.routing.model.response.ErrorResponse;
+import ua.everybuy.routing.model.response.StatusResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
