--- src/main/java/ua/everybuy/routing/model/model/response/ResponseMarker.java
@@ -1,4 +1,0 @@
-package ua.everybuy.routing.model.model.response;
-
-public interface ResponseMarker {
-}
