+++ src/main/java/ua/everybuy/routing/model/response/FullNameResponse.java
@@ -1,6 +1,5 @@
-package ua.everybuy.routing.model.model.response;
+package ua.everybuy.routing.model.response;
 
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
