+++ src/main/java/ua/everybuy/routing/model/response/ResponseMarker.java
@@ -1,0 +1,4 @@
+package ua.everybuy.routing.model.response;
+
+public interface ResponseMarker {
+}
