+++ src/main/java/ua/everybuy/routing/model/response/StatusResponse.java
@@ -1,4 +1,4 @@
-package ua.everybuy.routing.model.model.response;
+package ua.everybuy.routing.model.response;
 
 import lombok.*;
